minor code rearrange
