+++ lib/util.js
@@ -5,38 +5,8 @@ const fs = require('fs')
 const path = require('path')
 const { spawn } = require('child_process')
 const which = require('which')
-const pump = require('pump')
-const pumpify = require('pumpify')
-const split = require('split2')
-const through = require('through2')
-const browserify = require('browserify')
-const multistream = require('multistream')
-const concat = require('concat-stream')
-const eos = require('end-of-stream')
 const debug = require('debug')('0x')
-const launch = require('opn')
-const { promisify } = require('util')
 const profLogConvert = require('./prof-log-convert')
-const html = require('../visualizer/html')
-
-const phases = [
-  (frames) => frames,
-  (frames) => {
-    var moduleRunMain = frames.find(({name}) => /Module.runMain module.js/.test(name))
-    if (moduleRunMain) {
-      const startupIndex = frames.findIndex(({name}) => /startup bootstrap_node/.test(name))
-      if (startupIndex !== -1) frames.splice(startupIndex, 1)
-      return frames
-    }
-    var startup = frames.find(({name}) => /startup bootstrap_node/.test(name))
-    if (startup) return false
-    return frames
-  },
-  (frames) => {
-    if (frames.find(({name}) => /startup bootstrap_node/.test(name))) return false
-    return frames
-  }
-]
 
 module.exports = {
   determineOutputDir: determineOutputDir,
@@ -44,10 +14,7 @@ module.exports = {
   tidy: tidy,
   pathTo: pathTo,
   isSudo: isSudo,
-  noop: noop,
-  render,
-  createBundle,
-  phases
+  noop: noop
 }
 
 function determineOutputDir (args, proc) {
@@ -57,22 +24,6 @@ function determineOutputDir (args, proc) {
   .replace('{name}', args.name)
 
   return path.resolve(args.workingDir, name)
-}
-
-function determineHtmlPath (args, {pid, folder}) {
-  if (args.name === '-') return '-'
-  var htmlPath = (args.outputHtml || ( 
-      `{outputDir}${path.sep}{name}.html`
-  )).replace('{pid}', pid || 'UNKNOWN_PID')
-  .replace('{timestamp}', Date.now())
-  .replace('{outputDir}', folder)
-  .replace('{cwd}', args.workingDir)
-  .replace('{name}', args.name)
-
-  if (path.isAbsolute(htmlPath) === false) {
-    htmlPath = path.join(args.workingDir, htmlPath)
-  }
-  return htmlPath
 }
 
 
@@ -86,46 +37,6 @@ function ensureDirExists (path) {
       throw e
     }
   }
-}
-
-function createBundle () {
-  return new Promise((resolve, reject) => {
-    pump(
-      multistream([
-        browserify({standalone: 'hsl'}).add(require.resolve('hsl-to-rgb-for-reals')).bundle(),
-        browserify({standalone: 'fg'}).add(require.resolve('d3-fg')).bundle(),
-        browserify({standalone: 'ticksToTree'}).add(path.resolve(__dirname, 'ticks-to-tree')).bundle(),
-        browserify({standalone: 'visualizer'}).add(path.resolve(__dirname, '..', 'visualizer')).bundle()
-      ]),
-      concat(resolve),
-      (err) => err && reject(err)
-    )
-  })
-}
-const ticksToTree = require('./ticks-to-tree')
-async function render (args, {ticks, inlined, pid, folder}) {
-  const { name, title, kernelTracing } = args
-  debug('converted stacks to intermediate format')
-  const trees = ticksToTree(ticks, args.mapFrames, inlined)
-
-  const script = `
-    ${await createBundle()}
-    visualizer(${JSON.stringify(trees)}, ${JSON.stringify({title, kernelTracing})})
-  `
-
-  const htmlPath = determineHtmlPath(args, {pid, folder})
-  const opts = {
-    title,
-    name,
-    script,
-    htmlPath,
-    dir: folder,
-    stdout: name === '-' || htmlPath === '-'
-  }
-
-  await html(opts)
-  debug('done rendering')
-  return htmlPath === '-' ? null : `file://${htmlPath}`
 }
 
 function tidy (args) {
